tests/wpt/tests/navigation-api/currententrychange-event/navigation-back-forward-cross-doc.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
promise_test(async t => {
  await new Promise(resolve => window.onload = resolve);
  i.contentWindow.navigation.navigate("/common/blank.html?1");
  await new Promise(resolve => i.onload = resolve);
  assert_equals(i.contentWindow.navigation.entries().length, 2);

  i.contentWindow.navigation.oncurrententrychange = t.unreached_func("currententrychange should not fire for cross-document navigations");
  i.contentWindow.navigation.back();
  await new Promise(resolve => i.onload = resolve);

  i.contentWindow.navigation.oncurrententrychange = t.unreached_func("currententrychange should not fire for cross-document navigations");
  i.contentWindow.navigation.forward();
  await new Promise(resolve => i.onload = resolve);
}, "currententrychange does not fire for cross-document navigation.back() and navigation.forward()");
</script>
